Move specific part to specific mod.
