+++ src/query/storages/parquet/src/lib.rs
@@ -24,7 +24,7 @@ mod utils;
 
 pub use parquet2::Parquet2Table;
 pub use parquet_part::ParquetPart;
-pub use parquet_part::ParquetRSRowGroupPart;
 pub use parquet_rs::ParquetRSPruner;
 pub use parquet_rs::ParquetRSReader;
+pub use parquet_rs::ParquetRSRowGroupPart;
 pub use parquet_rs::ParquetRSTable;
