+++ src/query/storages/parquet/src/parquet2/mod.rs
@@ -14,9 +14,11 @@
 
 mod parquet_reader;
 mod parquet_table;
+mod partition;
 mod processors;
 mod projection;
 mod pruning;
 mod statistics;
 
 pub use parquet_table::Parquet2Table;
+pub use partition::Parquet2RowGroupPart;
