+++ src/query/storages/parquet/src/parquet_rs/mod.rs
@@ -14,6 +14,7 @@
 
 mod parquet_reader;
 mod parquet_table;
+mod partition;
 mod pruning;
 mod row_group;
 mod source;
@@ -21,4 +22,5 @@ mod statistics;
 
 pub use parquet_reader::ParquetRSReader;
 pub use parquet_table::ParquetRSTable;
+pub use partition::ParquetRSRowGroupPart;
 pub use pruning::ParquetRSPruner;
